Rename remaining components to use new terminology
